Move client and schema
