+++ src/apps/shared/HttpClient/client.ts
@@ -1,6 +1,8 @@
 import createClient, { Middleware } from 'openapi-fetch';
 import { paths } from './schema';
 import { getNewApiHeaders } from '../../../apps/main/auth/service';
+import { logger } from '../logger/logger';
+import { onUserUnauthorized } from './background-process-clients';
 
 const middleware: Middleware = {
   async onRequest({ request }) {
@@ -9,6 +11,12 @@ const middleware: Middleware = {
       request.headers.set(key, value);
     });
     return request;
+  },
+  onResponse({ response }) {
+    if (response.status === 401) {
+      logger.warn({ msg: 'Request unauthorized' });
+      onUserUnauthorized();
+    }
   },
 };
